refactor: remove test ignore hacks and increase coverage
