+++ packages/slice/__tests__/android/jest.config.js
@@ -1,3 +1,3 @@
 const jestConfigurator = require("@times-components/jest-configurator").default;
 
-module.exports = jestConfigurator("android", __dirname, ["responsive.web.js"]);
+module.exports = jestConfigurator("android", __dirname);
